request/response structs -> messages.go
